requirements.txt
anthropic>=0.40.0
flask>=3.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pytest>=8.0.0
